pyproject.toml
[tool.poetry]
name = "integritastecnica"
version = "0.1.0"
description = ""
authors = ["josejuniorpg <yunipgx@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
django-extensions = "^3.2.3"
django-model-utils = "^5.0.0"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
